refactor: extração de funções
